+++ packages/connectors/hull-zapier/test/integration/receive-subscription.test.js
@@ -13,7 +13,7 @@ it("Receive New Subscription", () => {
       handlerType: handlers.incomingRequestHandler,
       connector: {
         private_settings: {
-          subscriptions: []
+          triggers: []
         }
       },
       usersSegments: [],
@@ -26,7 +26,9 @@ it("Receive New Subscription", () => {
               "url": "https://hooks.zapier.com/hooks/standard/1/1/",
               "action": "entered_segment",
               "entityType": "user",
-              "inputData": {}
+              "inputData": {
+                "user_segments": [ "segment_1" ]
+              }
             }
           );
       },
@@ -44,12 +46,14 @@ it("Receive New Subscription", () => {
         ["GET", "/api/v1/app", {}, {}],
         ["PUT", "/api/v1/9993743b22d60dd829001999", {}, {
           "private_settings": {
-            "subscriptions": [
+            "triggers": [
               {
-                "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-                "action": "entered_segment",
-                "entityType": "user",
-                "inputData": {}
+                "serviceAction": {
+                  "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+                },
+                "inputData": {
+                  "entered_user_segments": [ "segment_1" ]
+                }
               }
             ]
           },
@@ -67,12 +71,14 @@ it("Receive New Subscription And Merge With Existing Subscriptions", () => {
       handlerType: handlers.incomingRequestHandler,
       connector: {
         private_settings: {
-          subscriptions: [
+          triggers: [
             {
-              "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-              "action": "entered_segment",
-              "entityType": "user",
-              "inputData": {}
+              "serviceAction": {
+                "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+              },
+              "inputData": {
+                "entered_user_segments": [ "segment_1" ]
+              }
             }
           ]
         }
@@ -87,7 +93,9 @@ it("Receive New Subscription And Merge With Existing Subscriptions", () => {
               "url": "https://hooks.zapier.com/hooks/standard/1/2/",
               "action": "entered_segment",
               "entityType": "user",
-              "inputData": {}
+              "inputData": {
+                "user_segments": [ "segment_2" ]
+              }
             }
           );
       },
@@ -105,18 +113,22 @@ it("Receive New Subscription And Merge With Existing Subscriptions", () => {
         ["GET", "/api/v1/app", {}, {}],
         ["PUT", "/api/v1/9993743b22d60dd829001999", {}, {
           "private_settings": {
-            "subscriptions": [
+            "triggers": [
               {
-                "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-                "action": "entered_segment",
-                "entityType": "user",
-                "inputData": {}
+                "serviceAction": {
+                  "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+                },
+                "inputData": {
+                  "entered_user_segments": [ "segment_1" ]
+                }
               },
               {
-                "url": "https://hooks.zapier.com/hooks/standard/1/2/",
-                "action": "entered_segment",
-                "entityType": "user",
-                "inputData": {}
+                "serviceAction": {
+                  "webhook": "https://hooks.zapier.com/hooks/standard/1/2/"
+                },
+                "inputData": {
+                  "entered_user_segments": [ "segment_2" ]
+                }
               }
             ]
           },
@@ -133,12 +145,14 @@ it("Receive New Subscription And Unable To Merge With Existing Subscriptions", (
       handlerType: handlers.incomingRequestHandler,
       connector: {
         private_settings: {
-          subscriptions: [
+          triggers: [
             {
-              "url": "https://hooks.zapier.com/hooks/standard/1/1/",
-              "action": "entered_segment",
-              "entityType": "user",
-              "inputData": {}
+              "serviceAction": {
+                "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+              },
+              "inputData": {
+                "entered_user_segments": [ "segment_1" ]
+              }
             }
           ]
         }
@@ -153,7 +167,9 @@ it("Receive New Subscription And Unable To Merge With Existing Subscriptions", (
               "url": "https://hooks.zapier.com/hooks/standard/1/1/",
               "action": "entered_segment",
               "entityType": "user",
-              "inputData": {}
+              "inputData": {
+                "user_segments": [ "segment_2" ]
+              }
             }
           );
       },
@@ -167,6 +183,69 @@ it("Receive New Subscription And Unable To Merge With Existing Subscriptions", (
       platformApiCalls: [["GET", "/api/v1/app", {}, {}],
         ["GET", "/api/v1/users_segments?shipId=9993743b22d60dd829001999", { "shipId": "9993743b22d60dd829001999" }, {}],
         ["GET", "/api/v1/accounts_segments?shipId=9993743b22d60dd829001999", { "shipId": "9993743b22d60dd829001999" }, {}]
+      ]
+    };
+  });
+});
+
+it("Receive New User Attribute Updated Subscription", () => {
+  return testScenario({ connectorConfig }, ({ handlers, nock, expect }) => {
+    return {
+      handlerType: handlers.incomingRequestHandler,
+      connector: {
+        private_settings: {
+          triggers: []
+        }
+      },
+      usersSegments: [],
+      accountsSegments: [],
+      externalIncomingRequest: ({ superagent, connectorUrl, plainCredentials }) => {
+        return superagent
+          .post(`${connectorUrl}/subscribe?ship=${plainCredentials.ship}&organization=${plainCredentials.organization}&secret=1234`)
+          .send(
+            {
+              "url": "https://hooks.zapier.com/hooks/standard/1/1/",
+              "action": "attribute_updated",
+              "entityType": "user",
+              "inputData": {
+                "user_segments": [ "segment_1" ],
+                "account_segments": [ "account_segment_1" ],
+                "user_attributes": [ "pipedrive/department" ],
+                "account_attributes": [ "pipedrive/industry", "num_employees" ]
+              }
+            }
+          );
+      },
+      response: expect.whatever(),
+      logs: [
+        ["info", "incoming.job.start", {}, { "jobName": "Incoming Data", "type": "webpayload" }],
+        ["info", "incoming.job.success", {}, { "jobName": "Incoming Data", "type": "webpayload" }]
+      ],
+      firehoseEvents:[],
+      metrics: [["increment", "connector.request", 1,]],
+      platformApiCalls: [
+        ["GET", "/api/v1/app", {}, {}],
+        ["GET", "/api/v1/users_segments?shipId=9993743b22d60dd829001999", { "shipId": "9993743b22d60dd829001999" }, {}],
+        ["GET", "/api/v1/accounts_segments?shipId=9993743b22d60dd829001999", { "shipId": "9993743b22d60dd829001999" }, {}],
+        ["GET", "/api/v1/app", {}, {}],
+        ["PUT", "/api/v1/9993743b22d60dd829001999", {}, {
+          "private_settings": {
+            "triggers": [
+              {
+                "serviceAction": {
+                  "webhook": "https://hooks.zapier.com/hooks/standard/1/1/"
+                },
+                "inputData": {
+                  "user_segments": [ "segment_1" ],
+                  "account_segments": [ "account_segment_1" ],
+                  "user_attribute_updated": [ "pipedrive/department" ],
+                  "account_attribute_updated": [ "pipedrive/industry", "num_employees" ]
+                }
+              }
+            ]
+          },
+          "refresh_status": false
+        }]
       ]
     };
   });